requirements.txt
streamlit
python-dotenv
google-generativeai
#pip install -r requirements.txt